Improve code for data processing and testing models
This commit implements missed functionality alongside improving the current ones. The use of MLFlow is expanded, allowing Random Forest model instances to be saved, loaded, and tested with Wandb and MLFlow tools. Code has also been added to split the dataset for training and testing purposes. It also implements utility tests for checking the 'row count' and 'price range' of the data. Moreover, updates made in the dependencies of conda.yml file.